Apple2 Done

+++ Infrastructure/Presentation/CoverController.cs
@@ -1,5 +1,6 @@
 ﻿using Apple1_Services.Abstractions;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using Shared;
 using System;
 using System.Collections.Generic;
@@ -11,8 +12,15 @@ namespace Presentation
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class CoverController(IServiceManager serviceManager) : ControllerBase
+    public class CoverController : ControllerBase
     {
+        private readonly string _adminPassword;
+        private readonly IServiceManager serviceManager;
+        public CoverController(IOptions<AdminSettings> adminSettingsOptions, IServiceManager service)
+        {
+            _adminPassword = adminSettingsOptions.Value.Password;
+            serviceManager = service;
+        }
         [HttpGet]
         public async Task<IActionResult> GetAllCovers()
         {
@@ -23,7 +31,7 @@ namespace Presentation
 
         [HttpGet("{Name}")]
 
-        public async Task<IActionResult> GetCoverByName([FromQuery] string name)
+        public async Task<IActionResult> GetCoverByName(string name)
         {
             var cover = await serviceManager.CoverService.GetCoverByNameAsync(name);
             if (cover == null) return NotFound($"Cover with name '{name}' not found.");
@@ -31,24 +39,34 @@ namespace Presentation
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateCover([FromBody] AddCoverResultDto coverDto)
+        public async Task<IActionResult> CreateCover([FromBody] AddCoverResultDto coverDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
+            if (ModelState.IsValid == false)
+            {
+                return BadRequest(ModelState);
+            }
             if (coverDto == null) return BadRequest("Cover data is null.");
             await serviceManager.CoverService.CreateCoverAsync(coverDto);
             return Ok(coverDto);
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateCover([FromBody] AddCoverResultDto coverDto)
+        public async Task<IActionResult> UpdateCover([FromBody] CoverResultDto coverDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             if (coverDto == null) return BadRequest("Cover data is null.");
             await serviceManager.CoverService.UpdateCoverAsync(coverDto);
             return Ok(coverDto);
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteCover(int id)
+        public async Task<IActionResult> DeleteCover(int id, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             await serviceManager.CoverService.DeleteCoverAsync(id);
             return Ok($"Cover with ID '{id}' has been deleted.");
         }
