+++ Infrastructure/Presentation/HeadPhoneController.cs
@@ -1,5 +1,6 @@
 ﻿using Apple1_Services.Abstractions;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using Shared;
 using System;
 using System.Collections.Generic;
@@ -11,8 +12,16 @@ namespace Presentation
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class HeadPhoneController(IServiceManager serviceManager) : ControllerBase
+    public class HeadPhoneController : ControllerBase
     {
+
+        private readonly string _adminPassword;
+        private readonly IServiceManager serviceManager;
+        public HeadPhoneController(IOptions<AdminSettings> adminSettingsOptions, IServiceManager service)
+        {
+            _adminPassword = adminSettingsOptions.Value.Password;
+            serviceManager = service;
+        }
         [HttpGet]
         public async Task<IActionResult> GetAllHeadPhones()
         {
@@ -22,7 +31,7 @@ namespace Presentation
         }
 
         [HttpGet("{Name}")]
-        public async Task<IActionResult> GetHeadPhoneByName([FromQuery] string name)
+        public async Task<IActionResult> GetHeadPhoneByName(string name)
         {
             var headPhone = await serviceManager.HeadPhoneService.GetHeadPhonesByNameAsync(name);
             if (headPhone == null) return NotFound($"Head phone with name '{name}' not found.");
@@ -30,24 +39,34 @@ namespace Presentation
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateHeadPhone([FromBody] AddHeadPhoneResultDto headPhoneDto)
+        public async Task<IActionResult> CreateHeadPhone([FromBody] AddHeadPhoneResultDto headPhoneDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
+            if (ModelState.IsValid == false)
+            {
+                return BadRequest(ModelState);
+            }
             if (headPhoneDto == null) return BadRequest("Head phone data is null.");
             await serviceManager.HeadPhoneService.CreateHeadPhonesAsync(headPhoneDto);
             return Ok(headPhoneDto);
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateHeadPhone([FromBody] AddHeadPhoneResultDto headPhoneDto)
+        public async Task<IActionResult> UpdateHeadPhone([FromBody] HeadPhoneResultDto headPhoneDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             if (headPhoneDto == null) return BadRequest("Head phone data is null.");
             await serviceManager.HeadPhoneService.UpdateHeadPhonesAsync(headPhoneDto);
             return Ok(headPhoneDto);
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteHeadPhone(int id)
+        public async Task<IActionResult> DeleteHeadPhone(int id, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             await serviceManager.HeadPhoneService.DeleteHeadPhonesAsync(id);
             return Ok($"Head phone with ID '{id}' has been deleted.");
         }
